[Refactor] #120 - ScanView 하위뷰들 정리중
